Consistently use u32 for number of bits

See merge request surfer-project/surfer!808

+++ libsurfer/src/translation/instruction_translators.rs
@@ -16,7 +16,7 @@ impl BasicTranslator<VarId, ScopeId> for InstructionTranslator {
         self.name.clone()
     }
 
-    fn basic_translate(&self, num_bits: u64, value: &VariableValue) -> (String, ValueKind) {
+    fn basic_translate(&self, num_bits: u32, value: &VariableValue) -> (String, ValueKind) {
         let u64_value = match value {
             VariableValue::BigUint(v) => v.to_u64_digits().last().copied(),
             VariableValue::String(s) => match check_vector_variable(s) {
